Minor enhancements for the line graph.
In this commit, a number of defaults have been used to make it easier to continue development on the visualization. Issue #2 has been opened to further consider which elements need attention, either by requiring the data sent to include this data, or by inferring the parameters. In either case, some sane defaults should be used in order to prevent issues (i.e. issues will occur when a wrong type of data type is used for an axis in a visualization)